refactor: Improve examples & tests structure
All examples now share next structure,

```
<example>/
- README.rst
- src/
  - <example>/
    - openapi.yaml
    - *.py
```

This results in moving all examples tests into `tests/examples/<example>/` and
moving `rororo` tests into `tests/rororo/`.

Also define better mypy config to properly found `rororo` library for
examples code.
