+++ examples/hobotnica/src/hobotnica/openapi.yaml
