+++ examples/hobotnica/src/hobotnica/views.py
