+++ examples/petstore/src/petstore/petstore-expanded.yaml
