+++ examples/petstore/src/petstore/shortcuts.py
