+++ mypy.ini
@@ -9,8 +9,10 @@ disallow_untyped_decorators = True
 disallow_untyped_defs = True
 follow_imports = normal
 follow_imports_for_stubs = True
-ignore_missing_imports = True
+ignore_missing_imports = False
 namespace_packages = True
+mypy_path = ./src/
+python_executable = ./.venv/bin/python3
 show_column_numbers = True
 show_error_codes = True
 warn_redundant_casts = True
@@ -28,3 +30,24 @@ ignore_errors = True
 
 [mypy-*.tests.*]
 ignore_errors = True
+
+[mypy-aioredis.*]
+ignore_missing_imports = True
+
+[mypy-email_validator.*]
+ignore_missing_imports = True
+
+[mypy-environ.*]
+ignore_missing_imports = True
+
+[mypy-isodate.*]
+ignore_missing_imports = True
+
+[mypy-jsonschema.*]
+ignore_missing_imports = True
+
+[mypy-openapi_core.*]
+ignore_missing_imports = True
+
+[mypy-openapi_schema_validator.*]
+ignore_missing_imports = True
