+++ pyproject.toml
@@ -3,8 +3,8 @@ version_type = "semver"
 version_files = [
   "pyproject.toml",
   "src/rororo/__init__.py",
-  "tests/openapi.json",
-  "tests/openapi.yaml"
+  "tests/rororo/openapi.json",
+  "tests/rororo/openapi.yaml"
 ]
 
 [tool.black]
@@ -33,7 +33,7 @@ homepage = "https://igordavydenko.com/projects/#rororo"
 repository = "https://github.com/playpauseandstop/rororo"
 documentation = "https://rororo.readthedocs.io/"
 packages = [
-  {include = "rororo", from = "src"},
+  {include = "rororo", from = "src/"},
 ]
 keywords = ["aiohttp", "aiohttp.web", "oas", "openapi", "openapi3", "helpers", "utilities"]
 classifiers = [
@@ -67,23 +67,17 @@ typing-extensions = {python = "<3.8", version = "^3.7.4"}
 aioredis = "^1.3.1"
 async_generator = {python = "~3.6", version = "^1.10"}
 coverage = {extras = ["toml"], version = "^5.3"}
-pytest = "^6.1.0"
+pytest = "^6.1.2"
 pytest-aiohttp = "^0.3.0"
-pytest-cov = "^2.7"
+pytest-cov = "^2.10.1"
 pytest-env = "^0.6.2"
 
 [tool.poetry.urls]
 "Bug Tracker" = "https://github.com/playpauseandstop/rororo/issues"
 
 [tool.pytest.ini_options]
-minversion = "6.0.1"
-testpaths = [
-  "examples/hobotnica/tests.py",
-  "examples/petstore/tests.py",
-  "examples/simulations/tests.py",
-  "examples/todobackend/tests.py",
-  "tests/",
-]
+minversion = "6.1.2"
+testpaths = ["tests/"]
 addopts = "--cov --no-cov-on-fail"
 env = """
 LEVEL=test
@@ -109,7 +103,6 @@ passenv =
   PYTEST_ADDOPTS
   REDIS_URL
 setenv =
-  PYTHONPATH=.
   USER=playpauseandstop
 skip_install = True
 whitelist_externals = poetry
@@ -123,3 +116,7 @@ commands_pre =
   poetry install
   poetry run python -m pip install attrs==19.3.0 openapi-core==0.13.3
 """
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
