+++ tests/rororo/data/blank.png
