+++ tests/rororo/invalid-openapi.json
