+++ tests/rororo/invalid-openapi.yaml
