+++ tests/rororo/test_openapi_core_data.py
