+++ tests/rororo/test_openapi_exceptions.py
